url 파라미터 적용
