+++ frontend/FEweek06HW/week06hw/package.json
@@ -13,6 +13,7 @@
     "axios": "^1.10.0",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
+    "react-router-dom": "^7.6.3",
     "styled-components": "^6.1.19"
   },
   "devDependencies": {
